add: properties menu in the right for selected Item
